Code clean up, remove old google analyitcs references
