+++ client/lib/analytics/ad-tracking/index.js
@@ -1,12 +1,6 @@
 export { adTrackRegistration } from './ad-track-registration';
 export { adTrackSignupComplete } from './ad-track-signup-complete';
 export { adTrackSignupStart } from './ad-track-signup-start';
-export {
-	fireGoogleAnalyticsEvent,
-	fireGoogleAnalyticsPageView,
-	getGoogleAnalyticsDefaultConfig,
-	setupGoogleAnalyticsGtag,
-} from './google-analytics';
 export { recordAddToCart } from './record-add-to-cart';
 export { recordAliasInFloodlight } from './record-alias-in-floodlight';
 export { recordOrder } from './record-order';
